Adding _BaseKPCov and modifying KPCovR/KPCovC to inherit from it
